requirements.txt
pyserial~=3.5
schedule~=1.2.2
requests~=2.32.3